+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 ## Enhancements
 * Adds the `ProviderType` field to `AdminSAMLSetting` and `AdminSAMLSettingsUpdateOptions` to support the `provider-type` SAML setting by @skj-skj [#1303](https://github.com/hashicorp/go-tfe/pull/1303)
-* Adds `AdminScimSetting`, `AdminScimSettingUpdateOptions`, and `AdminSettings.SCIM` to support managing site-level SCIM settings by @skj-skj [#1307](https://github.com/hashicorp/go-tfe/pull/1307)
+* Adds `AdminSCIMSetting`, `AdminSCIMSettingUpdateOptions`, and `AdminSettings.SCIM` to support managing site-level SCIM settings by @skj-skj [#1307](https://github.com/hashicorp/go-tfe/pull/1307)
 
 # v1.103.0